get rid of org detail
