+++ console/src/main/java/org/georchestra/console/ws/backoffice/orgs/OrgsController.java
@@ -24,7 +24,6 @@ import org.georchestra.console.dao.AdvancedDelegationDao;
 import org.georchestra.console.dao.DelegationDao;
 import org.georchestra.console.ds.OrgsDao;
 import org.georchestra.console.dto.orgs.Org;
-import org.georchestra.console.dto.orgs.OrgDetail;
 import org.georchestra.console.dto.orgs.OrgExt;
 import org.georchestra.console.model.DelegationEntry;
 import org.georchestra.console.ws.backoffice.utils.ResponseUtil;
@@ -163,9 +162,7 @@ public class OrgsController {
 
         Org org = this.orgDao.findByCommonName(cn);
         OrgExt orgExt = this.orgDao.findExtById(cn);
-        OrgDetail orgDetail = this.orgDao.findDetailById(cn);
         org.setOrgExt(orgExt);
-        org.setOrgDetail(orgDetail);
         return org;
     }
 
@@ -229,14 +226,11 @@ public class OrgsController {
         // Retrieve current orgs state from ldap
         Org org = this.orgDao.findByCommonName(commonName);
         OrgExt orgExt = this.orgDao.findExtById(commonName);
-        OrgDetail orgDetail = this.orgDao.findDetailById(commonName);
 
         // Update org and orgExt fields
         this.updateFromRequest(org, json);
         orgExt.setId(org.getId());
-        orgDetail.setId(org.getId());
         this.updateFromRequest(orgExt, json);
-        this.updateFromRequest(orgDetail, json);
         // Persist changes to LDAP server
         this.orgDao.update(org);
 
@@ -248,12 +242,8 @@ public class OrgsController {
             }
         }
 
-
-
         this.orgDao.update(orgExt);
-        this.orgDao.update(orgDetail);
         org.setOrgExt(orgExt);
-        org.setOrgDetail(orgDetail);
         return org;
     }
 
@@ -296,18 +286,11 @@ public class OrgsController {
         orgExt.setId(org.getId());
         this.updateFromRequest(orgExt, json);
 
-        OrgDetail orgDetail = new OrgDetail();
-        orgDetail.setId(org.getId());
-        this.updateFromRequest(orgDetail, json);
-
         // Persist changes to LDAP server
         this.orgDao.insert(org);
         this.orgDao.insert(orgExt);
-        this.orgDao.insert(orgDetail);
 
         org.setOrgExt(orgExt);
-        org.setOrgDetail(orgDetail);
-
         return org;
     }
 
@@ -328,7 +311,6 @@ public class OrgsController {
         // delete entities in LDAP server
         this.orgDao.delete(this.orgDao.findExtById(commonName));
         this.orgDao.delete(this.orgDao.findByCommonName(commonName));
-        this.orgDao.delete(this.orgDao.findDetailById(commonName));
         ResponseUtil.writeSuccess(response);
     }
 
@@ -492,13 +474,6 @@ public class OrgsController {
         org.setPending(json.optBoolean(Org.JSON_PENDING));
     }
 
-    protected void updateFromRequest(OrgDetail org, JSONObject json) {
-        org.setUrl(json.optString(Org.JSON_URL));
-        org.setPending(json.optBoolean(Org.JSON_PENDING));
-        org.setUrl(json.optString(Org.JSON_URL));
-        org.setLogo(json.optString(Org.JSON_LOGO));
-    }
-
     /**
      * Update orgExt instance based on field found in json object.
      *
@@ -513,6 +488,8 @@ public class OrgsController {
         orgExt.setAddress(json.optString(OrgExt.JSON_ADDRESS));
         orgExt.setPending(json.optBoolean(Org.JSON_PENDING));
         orgExt.setDescription(json.optString(Org.JSON_DESCRIPTION));
+        orgExt.setUrl(json.optString(Org.JSON_URL));
+        orgExt.setLogo(json.optString(Org.JSON_LOGO));
     }
 
     /**
