+++ console/src/test/java/org/georchestra/console/ws/newaccount/NewAccountFormControllerTest.java
@@ -15,9 +15,7 @@ import org.georchestra.console.dto.Account;
 import org.georchestra.console.dto.AccountFactory;
 import org.georchestra.console.dto.orgs.AbstractOrg;
 import org.georchestra.console.dto.orgs.Org;
-import org.georchestra.console.dto.orgs.OrgDetail;
 import org.georchestra.console.dto.orgs.OrgExt;
-import org.georchestra.console.dto.ReferenceAware;
 import org.georchestra.console.mailservice.EmailFactory;
 import org.georchestra.console.ws.utils.PasswordUtils;
 import org.georchestra.console.ws.utils.Validation;
@@ -168,13 +166,11 @@ public class NewAccountFormControllerTest {
 
         assertTrue(ret.equals("welcomeNewUser"));
         ArgumentCaptor<AbstractOrg> orgCaptor = ArgumentCaptor.forClass(AbstractOrg.class);
-        verify(mockOrgDao, times(3)).insert(orgCaptor.capture());
+        verify(mockOrgDao, times(2)).insert(orgCaptor.capture());
         assertTrue(orgCaptor.getAllValues().get(0).isPending());
         assertTrue(orgCaptor.getAllValues().get(0).getClass().equals(Org.class));
         assertTrue(orgCaptor.getAllValues().get(1).isPending());
         assertTrue(orgCaptor.getAllValues().get(1).getClass().equals(OrgExt.class));
-        assertTrue(orgCaptor.getAllValues().get(2).isPending());
-        assertTrue(orgCaptor.getAllValues().get(2).getClass().equals(OrgDetail.class));
 
     }
 
@@ -188,13 +184,11 @@ public class NewAccountFormControllerTest {
 
         assertTrue(ret.equals("welcomeNewUser"));
         ArgumentCaptor<AbstractOrg> orgCaptor = ArgumentCaptor.forClass(AbstractOrg.class);
-        verify(mockOrgDao, times(3)).insert(orgCaptor.capture());
+        verify(mockOrgDao, times(2)).insert(orgCaptor.capture());
         assertFalse(orgCaptor.getAllValues().get(0).isPending());
         assertTrue(orgCaptor.getAllValues().get(0).getClass().equals(Org.class));
         assertFalse(orgCaptor.getAllValues().get(1).isPending());
         assertTrue(orgCaptor.getAllValues().get(1).getClass().equals(OrgExt.class));
-        assertFalse(orgCaptor.getAllValues().get(2).isPending());
-        assertTrue(orgCaptor.getAllValues().get(2).getClass().equals(OrgDetail.class));
     }
 
     /**
